requirements.txt
streamlit
torch
torch-geometric
pandas
numpy
opencv-python-headless
easyocr
pillow
pdf2image
python-docx
googletrans==4.0.0-rc1
transformers
sentence-transformers
scikit-learn